Structural Changes in project
So far I have moved the headers of the code to the folder ./inc/

#Changes in process
In order to work with external resources I will have to make a copy of
them in new folders. This way the project will be compact.

##[]Drivers
The place for the CMSIS and HAL drivers, fundamental parts of
the code to work with MCUs ARM C-M4.

##[]Middleware
is the place to save libraries buildt on top of the drivers that is
provided by manufacturers or other external contributors.

##[]Unity
Using Unity is fine since it is the right place for tests for now and
pretty illustrative. But I am going to include there all the files
needed instead of loading an external library.
* Unity source code
* No external library